+++ reconstruction/reconstruction_window.h
@@ -35,7 +35,7 @@ private:
     void load_b_table(void);
 private:
     std::auto_ptr<QTimer> timer;
-    std::vector<image::affine_transform<float> > motion_args;
+    std::vector<image::affine_transform<double> > motion_args;
     unsigned int progress;
     bool terminated;
     std::auto_ptr<std::future<void> > motion_correction_thread;
